refactor: 💡 move actions to index level
✅ Closes: https://hashicorp.atlassian.net/browse/ICU-12619
