+++ ui/admin/app/controllers/scopes/scope/auth-methods/auth-method/managed-groups/managed-group/index.js
@@ -1,7 +1,7 @@
 import Controller, { inject as controller } from '@ember/controller';
 
 export default class ScopesScopeAuthMethodsAuthMethodManagedGroupsManagedGroupIndexController extends Controller {
-  @controller('scopes/scope/auth-methods') authMethods;
-  @controller('scopes/scope/auth-methods/auth-method/managed-groups')
+  @controller('scopes/scope/auth-methods/index') authMethods;
+  @controller('scopes/scope/auth-methods/auth-method/managed-groups/index')
   managedGroups;
 }
