+++ ui/admin/app/controllers/scopes/scope/auth-methods/auth-method/managed-groups/managed-group/members.js
@@ -1,6 +1,6 @@
 import Controller, { inject as controller } from '@ember/controller';
 
 export default class ScopesScopeAuthMethodsAuthMethodManagedGroupsManagedGroupMembersController extends Controller {
-  @controller('scopes/scope/auth-methods/auth-method/managed-groups')
+  @controller('scopes/scope/auth-methods/auth-method/managed-groups/index')
   managedGroups;
 }
