+++ ui/admin/app/controllers/scopes/scope/auth-methods/index.js
@@ -4,8 +4,11 @@
  */
 
 import Controller from '@ember/controller';
+import { action } from '@ember/object';
 
 export default class ScopesScopeAuthMethodsIndexController extends Controller {
+  // =services
+
   // =attributes
 
   /**
@@ -14,5 +17,36 @@ export default class ScopesScopeAuthMethodsIndexController extends Controller {
    */
   get hasPrimaryAuthMethod() {
     return Boolean(this.scopeModel.primary_auth_method_id);
+  }
+
+  // =actions
+
+  /**
+   * Removes an item from array `property` at `index` on the
+   * passed `authMethod`.  This is used to manage entries in fragment array
+   * fields such as `signing_algorithms`.
+   * @param {AuthMethodModel} authMethod
+   * @param {string} property
+   * @param {number} index
+   */
+  @action
+  async removeItemByIndex(authMethod, property, index) {
+    const array = authMethod.get(property).filter((item, i) => i !== index);
+    authMethod.set(property, array);
+  }
+
+  /**
+   * Adds a string item to array `property` on the passed `authMethod`.
+   * This is used to manage entries in fragment OIDC string array fields such
+   * as `signing_algorithms`.
+   * @param {AuthMethodModel} authMethod
+   * @param {string} property
+   * @param {string} value
+   */
+  @action
+  async addStringItem(authMethod, property, value) {
+    const existingArray = authMethod[property] ?? [];
+    const array = [...existingArray, { value }];
+    authMethod.set(property, array);
   }
 }
