--- ui/admin/tests/unit/controllers/scopes/scope/auth-methods-test.js
@@ -1,11 +1,0 @@
-import { module, test } from 'qunit';
-import { setupTest } from 'admin/tests/helpers';
-
-module('Unit | Controller | scopes/scope/auth-methods', function (hooks) {
-  setupTest(hooks);
-
-  test('it exists', function (assert) {
-    let controller = this.owner.lookup('controller:scopes/scope/auth-methods');
-    assert.ok(controller);
-  });
-});
